Local Behat is working!
- Headless Chrome ran out of memory in Docker so running it locally instead of inside Lando.
- Refactored Behat custom features
- Check HSTS headers on Pantheon
