--- .ci/lando/behat
@@ -1,29 +1,0 @@
-#!/bin/bash
-
-# Set up variables
-export BEHAT_ADMIN_PASSWORD=behat-on-lando-rocks
-export BEHAT_ADMIN_EMAIL=andrew+local-ci@pantheon.io
-export BEHAT_ADMIN_USERNAME=pantheon-ci-testing-local
-
-export RELOCATED_WP_ADMIN=TRUE
-
-# Dynamically set Behat configuration parameters
-export BEHAT_PARAMS='{"extensions":{"Behat\\MinkExtension":{"base_url":"'${LANDO_INTERNAL_SITE_URL}'"},"PaulGibbs\\WordpressBehatExtension":{"site_url":"'${LANDO_INTERNAL_SITE_URL}'/wp","users":{"admin":{"username":"'$BEHAT_ADMIN_USERNAME'","password":"'$BEHAT_ADMIN_PASSWORD'"}},"wpcli":{"binary":"wp"}}}}'
-
-# Kill existing Chrome instances
-pkill -f "(chrome)?(--headless)"
-
-# Start headless Chrome if needed
-if [[ $(jobs | grep chrome) ]]
-then
-    jobs | grep chrome
-    echo -e "\n Headless Chrome is already running ..."
-else
-    echo -e "\n Starting Chrome in headless mode ..."
-    google-chrome --disable-gpu --headless --remote-debugging-address=0.0.0.0 --remote-debugging-port=9222 --no-sandbox </dev/null &>/dev/null &
-    jobs | grep chrome
-    echo -e "\n Chrome started in headless mode ..."
-fi
-
-# Run WordHat
-./vendor/bin/behat --config=tests/behat/behat-pantheon.yml --strict --colors --format-settings='{"paths": false}' --tags 'no_auth'
