--- .ci/lando/create-wp-admin-user
@@ -1,19 +1,0 @@
-#!/bin/bash
-
-# Set up variables
-export BEHAT_ADMIN_PASSWORD=behat-on-lando-rocks
-export BEHAT_ADMIN_EMAIL=andrew+local-ci@pantheon.io
-export BEHAT_ADMIN_USERNAME=pantheon-ci-testing-local
-
-# Check if an admin user with our desired username exists
-export BEHAT_ADMIN_USER_EXISTS=$(wp user list --login=${BEHAT_ADMIN_USERNAME} --format=count)
-
-# Create the admin user if needed
-if [[ "$BEHAT_ADMIN_USER_EXISTS" == "0" ]]
-then
-  echo -e "\nCreating the WordPress admin user..."
-  wp user create $BEHAT_ADMIN_USERNAME $BEHAT_ADMIN_EMAIL --user_pass=$BEHAT_ADMIN_PASSWORD --role=administrator
-  echo -e "\n WordPress user $BEHAT_ADMIN_USERNAME with the password $BEHAT_ADMIN_PASSWORD has been created."
-else
-  echo -e "\n Using the existing WordPress user $BEHAT_ADMIN_USERNAME."
-fi
