--- .ci/lando/delete-wp-admin-user
@@ -1,17 +1,0 @@
-#!/bin/bash
-
-# Set up variables
-export BEHAT_ADMIN_PASSWORD=behat-on-lando-rocks
-export BEHAT_ADMIN_EMAIL=andrew+local-ci@pantheon.io
-export BEHAT_ADMIN_USERNAME=pantheon-ci-testing-local
-
-# Check if an admin user with our desired username exists
-export BEHAT_ADMIN_USER_EXISTS=$(wp user list --login=${BEHAT_ADMIN_USERNAME} --format=count)
-
-# Create the admin user if needed
-if [[ "$BEHAT_ADMIN_USER_EXISTS" == "1" ]]
-then
-  echo -e "\nDeleting the WordPress admin user..."
-  wp user delete $BEHAT_ADMIN_USERNAME --yes
-  echo -e "\n WordPress user $BEHAT_ADMIN_USERNAME has been deleted."
-fi
