+++ .circleci/config.yml
@@ -35,7 +35,7 @@ defaults: &defaults
     ADMIN_USERNAME: admin
     TERM: dumb
 
-version: 2.1
+version: 2
 
 jobs:
   configure_env_vars:
