+++ composer.json
@@ -43,7 +43,7 @@
     "dealerdirect/phpcodesniffer-composer-installer": "^0.5.0",
     "dmore/behat-chrome-extension": "^1.3.0",
     "genesis/behat-fail-aid": "^2.1",
-    "paulgibbs/behat-wordpress-extension": "^3.2.0",
+    "paulgibbs/behat-wordpress-extension": "^3.3.0",
     "phpunit/phpunit": "^8.2.2",
     "roave/security-advisories": "dev-master",
     "squizlabs/php_codesniffer": "^3.4.2",
