+++ tests/behat/behat-pantheon.yml
@@ -12,6 +12,7 @@ default:
         - PaulGibbs\WordpressBehatExtension\Context\WordpressContext
         - PaulGibbs\WordpressBehatExtension\Context\EditPostContext
         - PaulGibbs\WordpressBehatExtension\Context\WidgetContext
+        - ataylorme\WordHatHelpers\Contexts\WordPressContext
         - ataylorme\WordHatHelpers\Contexts\PantheonContext
         - ataylorme\WordHatHelpers\Contexts\MiscContext
